Removed focus from world/map select
Since we have the numbers and letters set as hotkeys, they
would sometimes not work since focus is still on the dropdown
menu.

This would then change the world/map to a world/map that starts
with the number 2 if you were to change to the map editor by
tapping 2 for example.