Can view previous saved revisions of map
Still need to fix revision changing when focused and number key is pressed as well
as changing back to current revision if an older revision is selected.

Also, the revision changes won't work using simple-git. I will be submitting a PR
to the repo later with an enhancement that lets you checkout a specific file from a revision.