Implemented #55
There is a race condition currently though that needs to be resolved.
From what I've seen so far, it looks like this is the correct way to render the map.
It isn't CPU intensive (except for the initial render) and should be able to help make the game
playable for older machines.

The race condition occurs in init.jswhen socket.on("preData") completing and launching the game before
the secondary canvas has a chance to finish drawing the map. Therefore, when the game attempts to read
from the secondary canvas, it gets undefined instead and doesn't draw anything.

Once we fix the async problem for setting up the game, I think that this will work perfectly!